Move Reconciliation* classes out of MediaWiki namespace
ReconciliationService and its result ReconciliationItem move into a new
Reconciliation namespace; ReconciliationException moves into the
top-level extension namespace, since it’s interpreted as the base class
for all exceptions of the extension, and is in fact mostly thrown by
code in other namespaces.
